Factor out game statistics methods and helper methods into modules
