+++ lib/helper_methods.rb
@@ -1,0 +1,6 @@
+module HelperMethods
+
+
+
+
+end
