+++ lib/stat_tracker.rb
@@ -1,10 +1,14 @@
 require_relative './team'
 require_relative './game'
 require_relative './game_team'
+require_relative './helper_methods'
+require_relative './game_stats'
 require 'csv'
 require 'pry'
 
 class StatTracker
+  include GameStats
+  include HelperMethods
 
   attr_reader :teams, :games, :game_teams
 
@@ -32,37 +36,5 @@ class StatTracker
 
     StatTracker.new(teams, games, game_teams)
   end
-
-
-  def highest_total_score
-    highest_game = @games.max_by do |game|
-      game.away_goals + game.home_goals
-    end
-
-    highest_game.away_goals + highest_game.home_goals
-  end
-
-  def lowest_total_score
-    lowest_game = @games.min_by do |game|
-      game.away_goals + game.home_goals
-    end
-
-    lowest_game.away_goals + lowest_game.home_goals
-  end
-
-
-
-  def count_of_games_by_season
-    count = Hash.new(0)
-
-    @games.each do |game|
-      count[game.season.to_s] += 1
-    end
-
-    count
-  end
-
-
-
 
 end
